Added tests for disable SSL certificate verify in Http task config
